requirements.txt
python-telegram-bot==13.15
pyodbc
python-dotenv
urllib3==1.26.18
six
